fix: restore Google registration handler and improve OAuth URL construction
